Add files via upload
Online hotel booking has radically changed booking options and customer behavior. A significant number of hotel reservations are canceled due to cancellations or no-shows. Typical reasons for cancellations include changes in plans, scheduling conflicts, etc. This is often facilitated by the ability to do this for free or preferably at low cost, which is possible for hotel guests, but is less desirable and perhaps a decreasing revenue factor for hotels to manage. .

Can we predict if the customer will honor the reservation or cancel it?